DS-2269 - update collection methods to utilize new abstractapicollection, which dries up collection classes and adds limit
