+++ src/Program/ProgramListResponse.php
@@ -2,9 +2,13 @@
 
 namespace AllDigitalRewards\RewardStack\Program;
 
-use AllDigitalRewards\RewardStack\Common\Entity\Program;
+use AllDigitalRewards\RewardStack\Common\AbstractCollectionApiResponse;
+use AllDigitalRewards\RewardStack\Common\Entity\ProgramRetrieve;
 
-class ProgramListResponse extends Program
+class ProgramListResponse extends AbstractCollectionApiResponse
 {
-
+    protected function getEntityClass(): string
+    {
+        return ProgramRetrieve::class;
+    }
 }
