+++ src/Transaction/TransactionRequest.php
@@ -2,10 +2,11 @@
 
 namespace AllDigitalRewards\RewardStack\Transaction;
 
+use AllDigitalRewards\RewardStack\Common\AbstractApiCollectionRequest;
+use AllDigitalRewards\RewardStack\Common\CollectionFilterInterface;
 use AllDigitalRewards\RewardStack\Common\Entity\AbstractEntity;
-use AllDigitalRewards\RewardStack\Common\AbstractApiRequest;
 
-class TransactionRequest extends AbstractApiRequest
+class TransactionRequest extends AbstractApiCollectionRequest
 {
     /**
      * @var string
@@ -16,34 +17,21 @@ class TransactionRequest extends AbstractApiRequest
      */
     private $uniqueId;
 
-    /**
-     * @var int
-     */
-    private $page = 1;
-
-    protected $httpMethod = 'GET';
-
-    /**
-     * TransactionRequest constructor.
-     * @param string $programId
-     * @param string $uniqueId
-     * @param int $page
-     */
-    public function __construct(string $programId, string $uniqueId, int $page = 1)
-    {
+    public function __construct(
+        string $programId,
+        string $uniqueId,
+        int $page = 1,
+        int $limit = 30,
+        CollectionFilterInterface $filter = null
+    ) {
+        parent::__construct($page, $limit, $filter);
         $this->programId = $programId;
         $this->uniqueId = $uniqueId;
-        $this->page = $page;
     }
 
     public function getHttpEndpoint(): string
     {
-        return '/api/user/' . $this->uniqueId. '/transaction' ;
-    }
-
-    public function getQueryParams(): string
-    {
-        return "page=" . $this->page;
+        return "/api/program/{$this->programId}/participant/$this->uniqueId/transaction";
     }
 
     public function getResponseObject(): AbstractEntity
